+++ MacroPad/App.axaml
@@ -19,7 +19,6 @@
 	</Application.Resources>
 	
     <Application.DataTemplates>
-        <local:ViewLocator/>
     </Application.DataTemplates>
   
     <Application.Styles>
