+++ MacroPad/Controls/DeviceViewer.axaml
