+++ MacroPad/Controls/DeviceViewer.axaml.cs
