+++ MacroPad/Views/MainWindow.axaml
@@ -2,19 +2,31 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:vm="using:MacroPad.ViewModels"
         xmlns:views="using:MacroPad.Views"
+		xmlns:ui="using:FluentAvalonia.UI.Controls"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         x:Class="MacroPad.Views.MainWindow"
-        x:DataType="vm:MainWindowViewModel"
+        x:DataType="views:MainWindowViewModel"
         Icon="/Assets/avalonia-logo.ico"
         Title="MacroPad">
 
     <Design.DataContext>
-        <vm:MainWindowViewModel/>
+		<views:MainWindowViewModel/>
     </Design.DataContext>
-	<Panel>
-		<views:MainView/>
-	</Panel>
+	
+	<Grid Margin="4 0 4 4">
+		<Grid.ColumnDefinitions>
+			<ColumnDefinition Width="1*" MaxWidth="320" MinWidth="192"/>
+			<ColumnDefinition Width="4*"/>
+		</Grid.ColumnDefinitions>
+		<Border Classes="card" Padding="4">
+			<Grid RowDefinitions="*,32">
+				<ScrollViewer >
+				</ScrollViewer>
+				<Button Theme="{StaticResource TransparentButton}" VerticalAlignment="Center" HorizontalAlignment="Right" Grid.Row="1" Width="32" Height="32" Padding="0" ><ui:SymbolIcon Symbol="Settings" FontSize="16"/></Button>
+			</Grid>
+		</Border>
+	</Grid>
 
 </Window>
