+++ MacroPad/Views/MainWindowViewModel.cs
@@ -1,4 +1,6 @@
-﻿namespace MacroPad.ViewModels
+﻿using MacroPad.ViewModels;
+
+namespace MacroPad.Views
 {
     internal class MainWindowViewModel : ViewModelBase
     {
