+++ MacroPad/Views/Navigation/DeviceNav.axaml
@@ -1,0 +1,16 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="MacroPad.Views.Navigation.DeviceNav">
+	<Border Classes="card" Padding="4">
+		<Panel>
+			<WrapPanel VerticalAlignment="Top" HorizontalAlignment="Left">
+				<home:DeviceViewer x:Name="DeviceIcon" IsEnabled="False" Width="128" Height="128" />
+				<TextBlock Classes="h5" Margin="8 4 0 0" x:Name="DeviceName" TextWrapping="Wrap"/>
+			</WrapPanel>
+			<CheckBox Content="Enable Device" x:Name="EnableDevice" Margin="0 4 0 0" VerticalAlignment="Bottom" HorizontalAlignment="Right"/>
+		</Panel>
+	</Border>
+</UserControl>
