+++ MacroPad/Views/Navigation/DeviceNav.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace MacroPad.Views.Navigation;
+
+public partial class DeviceNav : UserControl
+{
+    public DeviceNav()
+    {
+        InitializeComponent();
+    }
+}
